ramips: ethernet: ralink: add disable-mac-learn
When mt7530 is used only as a way to reach ananother switch chip, there
is no need for mac learning. Also, if the stacked switch forces the
forwarding to happen in the CPU (as upstream rtl8365mb currently does),
the source mac address learning will prevent bridged ports to talk with
each other.

The new OF property to disable source mac address learning is
"mediatek,mac-learn-disable" and swconfig is "disable_mac_learn".

Signed-off-by: Luiz Angelo Daros de Luca <luizluca@gmail.com>